Modified   Project/BinarySerializerManaged/BufferStream.cs Modified   Project/CommonManaged/Utilities/Random.cs Modified   Project/NetworkingManaged/BaseSocket.cs Modified   Project/NetworkingManaged/ClientSocket.cs Modified   Project/NetworkingManaged/Constants.cs Modified   Project/NetworkingManaged/ServerSocket.cs Modified   Project/Tests/ClientNetworkingManagedTest/Program.cs Modified   Project/Tests/ServerNetworkingManagedTest/Program.cs Modified   Project/Binaries/AnalyticsManaged.dll Modified   Project/Binaries/ASCIISerializerManaged.dll Modified   Project/Binaries/BinarySerializerManaged.dll Modified   Project/Binaries/CommonAPIManaged.dll Modified   Project/Binaries/CommonManaged.dll Modified   Project/Binaries/DatabaseManaged.dll Modified   Project/Binaries/NetworkingManaged.NetCore.dll Modified   Project/Binaries/NetworkingManaged.NetFramework.dll

+++ Project/BinarySerializerManaged/BufferStream.cs
@@ -157,9 +157,9 @@ namespace GameFramework.BinarySerializer
 			Size += (uint)Data.Length;
 		}
 
-		public void WriteBytes(byte[] Data, int Index, int Length)
+		public void WriteBytes(byte[] Data, uint Index, uint Length)
 		{
-			stream.Write(Data, Index, Length);
+			stream.Write(Data, (int)Index, (int)Length);
 			Size += (uint)Data.Length;
 		}
